fix: finish C4996/C4244 warning burn-down via CrtCompat.h wrappers (#24)
Adds src/engine/CrtCompat.h with warning-clean portable wrappers (sv::StrCopy, sv::FOpen, sv::GetEnv, sv::StrError, sv::U8Path) and migrates all strncpy/fopen/getenv/strerror/fs::u8path call sites in src/, lab/, tests/. Fixes the last C4244 in Audio.cpp with an explicit static_cast. Closes backlog items 2 and 10.

+++ src/engine/Audio.cpp
@@ -233,7 +233,7 @@ SoundHandle Audio::postEventInternal(const char* eventName, const glm::vec3* pos
 
     // Cooldown check
     if (def.cooldownMs > 0) {
-        double now = ma_engine_get_time_in_milliseconds(m_engine);
+        double now = static_cast<double>(ma_engine_get_time_in_milliseconds(m_engine));
         if (now - def.lastPlayTime < (double)def.cooldownMs)
             return INVALID_SOUND;
         def.lastPlayTime = now;
